Slight design touchup and error handling.

+++ hackday-frontend/src/app/admin/page.tsx
@@ -2,6 +2,7 @@
 
 import Link from 'next/link';
 import React from 'react';
+import './style.css';
 
 type MonsterDTO = {
   monster_id: number;
@@ -12,7 +13,8 @@ type MonsterDTO = {
 
 async function addMonster(event: React.FormEvent<HTMLFormElement>) {
   event.preventDefault();
-  const reqBody: MonsterDTO = { monster_id: event.currentTarget.monid.value,
+  const reqBody: MonsterDTO = {
+    monster_id: event.currentTarget.monid.value,
     monster_name: event.currentTarget.monname.value,
     monster_group: event.currentTarget.group1.value,
     monster_second_group: event.currentTarget.group2.value
@@ -28,29 +30,35 @@ async function addMonster(event: React.FormEvent<HTMLFormElement>) {
 async function deleteMonster(event: React.FormEvent<HTMLFormElement>) {
   event.preventDefault();
   const id: string = event.currentTarget.monid.value;
-  const reqOptions = {
-    method: 'DELETE'
-  };
-  const response = await fetch(`http://localhost:8080/api/monsters/${id}`, reqOptions);
+  if(id) {
+    const reqOptions = {
+      method: 'DELETE'
+    };
+    const response = await fetch(`http://localhost:8080/api/monsters/${id}`, reqOptions);
+  }
 }
 
 export default function Admin() {
   return (
     <div>
-      <Link href={"/"}>Home</Link>
-      <h3>Add a monster</h3>
-      <form className="adminform" id="addform" onSubmit={e => addMonster(e)}>
-        <input type="text" name="monid" placeholder="Dex Number" />
-        <input type="text" name="monname" placeholder="Name" />
-        <input type="text" name="group1" placeholder="Egg Group 1" />
-        <input type="text" name="group2" placeholder="Egg Group 2" />
-        <button type="submit">submit</button>
-      </form>
-      <h3>Delete a monster</h3>
-      <form className="adminform" id="deleteform" onSubmit={e => deleteMonster(e)}>
-        <input type="text" name="monid" placeholder="Dex Number" />
-        <button type="submit">submit</button>
-      </form>
+      <header className="header">
+        <Link className="link" href={"/"}>Home</Link>
+      </header>
+      <div className="flexcontainer">
+        <form className="adminform" id="addform" onSubmit={e => addMonster(e)}>
+          <h3>Add a monster</h3>
+          <input type="text" className="adminselect" name="monid" placeholder="Dex Number" />
+          <input type="text" className="adminselect" name="monname" placeholder="Name" />
+          <input type="text" className="adminselect" name="group1" placeholder="Egg Group 1" />
+          <input type="text" className="adminselect" name="group2" placeholder="Egg Group 2" />
+          <button type="submit">submit</button>
+        </form>
+        <form className="adminform" id="deleteform" onSubmit={e => deleteMonster(e)}>
+          <h3>Delete a monster</h3>
+          <input type="text" className="adminselect" name="monid" placeholder="Dex Number" />
+          <button type="submit">submit</button>
+        </form>
+      </div>
     </div>
   )
 }
